requirements.txt
langchain_community
ipykernel
jupyter
PyMuPDF
tqdm
sentence_transformers
tiktoken
langchainhub
chromadb
langchain
langsmith
bs4
python-dotenv
langchain_core
langchain
langchain_ollama
pypdf
PyMuPDF
langchain_huggingface
ollama
matplotlib
numpy
pandas
scikit-learn
seaborn
imbalanced-learn
xgboost
unstructured
PyPDF2
typing_extensions
ragas
datasets
deepeval
accelerate
bitsandbytes
streamlit
fastapi[standard]